requirements.txt
django~=3.1.0
django-allauth~=0.42
django-analytical~=2.6.0
django-cors-headers~=3.5.0
django-cors-middleware~=1.5
django-dotenv~=1.4.2
django-countries~=6.1.3
django-exclusivebooleanfield==0.3.0 # Not updated for Django 3
django-filter~=2.3.0
django-getenv~=1.3.2
django-invitations~=1.9.3
django-lifecycle~=0.7.7
django-oauth-toolkit~=1.3.2
django-timezone-field~=4.0
django-widget-tweaks~=1.4.8
djangorestframework~=3.11.1
git+git://github.com/chibisov/drf-extensions.git@002a67cec4159cbd691580b6e7fa42a701f0d2dd#egg=drf-extensions
humanize
pendulum~=2.1.2
Pillow
psycopg2-binary
Pygments
python-dateutil~=2.8.1
pyyaml
requests
uritemplate
